feat: group db functions

+++ server/src/db/marknotes.ts
@@ -4,6 +4,7 @@ import { ObjectId } from "mongodb";
 
 const mongoCollections = require("../config/mongoCollections");
 const marknotes = mongoCollections.marknotes;
+const groups = require("./groups");
 
 /**
  * Inserts a marknote into the database.
@@ -12,7 +13,7 @@ const marknotes = mongoCollections.marknotes;
  * @param body Content of the note.
  * @returns The new marknote. Throws an error if failed.
  */
-const createMarknote = async (title: string, color: string, body: string) => {
+export const createMarknote = async (title: string, color: string, body: string) => {
   if (color.trim().length === 0) throw "createMarknote: must provide a color";
   if (!isHex(color)) throw `createMarknote: '${color}}' is not a valid hex code`
   const newMarknote: Marknote = {
@@ -23,6 +24,7 @@ const createMarknote = async (title: string, color: string, body: string) => {
     body: body,
     lastModified: Date.now(),
     favorited: false,
+    groups: []
   };
 
   const marknotesCollection = await marknotes();
@@ -38,7 +40,7 @@ const createMarknote = async (title: string, color: string, body: string) => {
  * Returns a list of all quicknotes in the database.
  * @returns List of marknotes.
  */
-const getAllMarknotes = async () => {
+export const getAllMarknotes = async () => {
   const marknotesCollection = await marknotes();
   const marknotesList = await marknotesCollection.find({}).toArray();
 
@@ -52,16 +54,12 @@ const getAllMarknotes = async () => {
 /**
  * Returns a single quicknote by its id.
  * @param id Target marknote id.
- * @returns The marknote object if found. Otherwise, throws an error.
+ * @returns The marknote object if found. Otherwise, null.
  */
-const getMarknoteById = async (id: string) => {
+export const getMarknoteById = async (id: string) => {
   const parsed_id = new ObjectId(id.trim());
   const marknotesCollection = await marknotes();
   const marknote = await marknotesCollection.findOne({ _id: parsed_id });
-
-  if (marknote === null) throw `getMarknoteById: No marknote found with id '${id}'`;
-  marknote._id = marknote._id.toString();
-
   return marknote;
 };
 
@@ -71,7 +69,7 @@ const getMarknoteById = async (id: string) => {
  * @param updatedMarknote The updated marknote information.
  * @returns The updated marknote if successful. Otherwise, throws an error.
  */
-const updateMarknoteById = async (id: string, updatedMarknote: Marknote) => {
+export const updateMarknoteById = async (id: string, updatedMarknote: Marknote) => {
   if (!isHex(updatedMarknote.color)) throw `updateMarknote: '${updatedMarknote.color}}' is not a valid hex code`
   const marknotesCollection = await marknotes();
   const parsed_id = new ObjectId(id.trim());
@@ -89,21 +87,67 @@ const updateMarknoteById = async (id: string, updatedMarknote: Marknote) => {
  * @param id Target marknote id.
  * @returns True if successfully deleted. Otherwise, throws an error.
  */
-const deleteMarknoteById = async (id: string) => {
+export const deleteMarknoteById = async (id: string) => {
   const marknotesCollection = await marknotes();
   const parsed_id = new ObjectId(id.trim());
+
+  // Check if note exists
+  const note = await getMarknoteById(id);
+  if (!note) {
+    throw `deleteMarknoteById: Could not find marknote with id '${id}'`;
+  }
+  
+  // Delete the note
   const deleteInfo = await marknotesCollection.deleteOne({ _id: parsed_id });
   if (deleteInfo.deletedCount === 0)
     throw "deleteMarknoteById: Failed to delete marknote";
+
+    // Remove from all groups
+    for (const group_id of note.groups) {
+      try {
+        await groups.removeFromGroup(group_id, note._id.toString(), note.type, true)
+      } catch (e) {
+        console.log(e)
+      }
+    }
+
   return true;
 };
 
-module.exports = {
-  createMarknote,
-  getAllMarknotes,
-  getMarknoteById,
-  updateMarknoteById,
-  deleteMarknoteById,
+/**
+ * Adds a group to the marknote's groups array.
+ * @param note_id The target note id.
+ * @param group_id The target group id.
+ * @returns The updated marknote if successful. Otherwise, throws an error.
+ */
+export const addGroupToMarknote = async (note_id: string, group_id: string) => {
+  const marknotesCollection = await marknotes();
+  const parsed_id = new ObjectId(note_id.trim());
+
+  const updateInfo = await marknotesCollection.updateOne(
+    { _id: parsed_id },
+    { $addToSet: { groups: group_id } }
+  );
+  if (!updateInfo.matchedCount && !updateInfo.modifiedCount)
+    throw `addGroupToMarknote: Failed to add group {'id': '${group_id}'}' to note {'type': 'marknote', 'id': '${note_id}'}`;
+  return await getMarknoteById(note_id.trim());
 };
 
-export {};
+/**
+ * Removes a group from the marknote's group array.
+ * @param note_id The target note id.
+ * @param group_id The target group id.
+ * @returns The updated marknote if successful. Otherwise, throws an error.
+ */
+export const removeGroupFromMarknote = async (note_id: string, group_id: string) => {
+  const marknotesCollection = await marknotes();
+  const parsed_id = new ObjectId(note_id.trim());
+
+  const updateInfo = await marknotesCollection.updateOne(
+    { _id: parsed_id },
+    { $pull: { groups: group_id } }
+  );
+  if (!updateInfo.matchedCount && !updateInfo.modifiedCount)
+    throw `removeGroupFromMarknote: Failed to remove group {'id': '${group_id}'}' from note {'type': 'marknote', 'id': '${note_id}'}`;
+  return await getMarknoteById(note_id.trim());
+}
